feat: ResultDetailList 추상화 컴포넌트 추가
